Adapted plotting code to new preprocessing
